@@ -8,7 +8,7 @@
 
   <PropertyGroup>
     <Description>PsSetup</Description>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <PropertyGroup>
