@@ -7,7 +7,7 @@
     <Description>Microsoft.IIS.Administration.Certificates Class Library</Description>
     <VersionPrefix>$(IISAdministrationVersion)</VersionPrefix>
     <Authors>Microsoft</Authors>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <AssemblyName>Microsoft.IIS.Administration.Certificates</AssemblyName>
     <PackageId>Microsoft.IIS.Administration.Certificates</PackageId>
   </PropertyGroup>
