@@ -50,7 +50,7 @@ namespace Microsoft.IIS.Administration.AccessManagement {
             // Renew the key
             string token = await _keyProvider.RenewToken(key);
 
-            return Created(Request.RequestUri.PathAndQuery,
+            return Created(Request.Path.ToString(),
                            AccessTokenHelper.ToJsonModel(new ApiToken() { Token=token, Key=key }));
         }
 
